refactor: added use-user-teams-state hook
