+++ common/client/fragments/user-teams-fragment/src/hooks/index.ts
@@ -1,0 +1,1 @@
+export * from './use-user-teams-state.hook'
